Fixed camera angle reading from both openmv and teensy. Tilt from goalie kinda working now

+++ include/data_source_camera_conicmirror.h
@@ -19,7 +19,7 @@ class DataSourceCameraConic : public DataSource{
         int yAngle, bAngle, yAngleFix, bAngleFix, yDist, bDist;
 
         int count = 0, unkn_counter;
-        byte xb, yb, xy, yy, true_xb, true_xy, true_yb, true_yy, calc_xb, calc_yb, calc_xy, calc_yy;
+        int xb, yb, xy, yy, true_xb, true_xy, true_yb, true_yy, calc_xb, calc_yb, calc_xy, calc_yy;
         bool data_received = false, start = false, end = false;
 
         int goalOrientation, pAtk, pDef;
